+++ platform/backend/src/services/agent-tool-assignment.ts
@@ -87,56 +87,6 @@ export async function assignToolToAgent(
   return null;
 }
 
-export async function assignAgentToolsFromLabels(
-  agentId: string,
-): Promise<void> {
-  // Get the agent and verify it's eligible for automatic tool assignment based on labels
-  const agent = await AgentModel.findById(agentId);
-  if (!agent) return;
-  if (agent.agentType !== "mcp_gateway") return;
-  if (agent.toolAssignmentMode !== "automatic") return;
-
-  // Fetch the agent's labels and determine which tools should be assigned based on those labels
-  const labels = await AgentLabelModel.getLabelsForAgent(agentId);
-
-  // For each label, find catalog items that match the label, then find tools associated with those catalog items, and build a set of desired tool IDs to be assigned to the agent
-  const desiredToolIdsSet = new Set<string>();
-  if (labels.length > 0) {
-    const catalogIds = await McpCatalogLabelModel.getCatalogIdsByLabels(labels);
-
-    if (catalogIds.length > 0) {
-      const toolIds = await ToolModel.getToolIdsByCatalogIds(catalogIds);
-      for (const toolId of toolIds) {
-        desiredToolIdsSet.add(toolId);
-      }
-    }
-  }
-
-  // Fetch the agent's assigned tool IDs and determine which tools to add/remove
-  const currentToolIds = await AgentToolModel.findToolIdsByAgent(agentId);
-  const currentToolIdsSet = new Set(currentToolIds);
-
-  const desiredToolIds = Array.from(desiredToolIdsSet);
-  const toInsert = desiredToolIds.filter((id) => !currentToolIdsSet.has(id));
-  const toDelete = currentToolIds.filter((id) => !desiredToolIdsSet.has(id));
-
-  // Assign new tools that are not currently assigned to the agent.
-  if (toInsert.length > 0) {
-    const entries = toInsert.map((toolId) => ({
-      agentId,
-      toolId,
-      credentialResolutionMode: "dynamic" as CredentialResolutionMode, // default to dynamic for automatically assigned tools
-    }));
-
-    await AgentToolModel.bulkCreate(entries);
-  }
-
-  // Remove tools that are currently assigned but not in the desired set.
-  if (toDelete.length > 0) {
-    await AgentToolModel.deleteManyForAgent(agentId, toDelete);
-  }
-}
-
 export async function validateAssignment(
   params: AgentToolAssignmentRequest,
 ): Promise<AgentToolAssignmentError | null> {
